New method finding correct "cpu" sting
Except of using the platform.architecture for the check, I'm using the platform.machine output to check for the cpu detection.
I'm not sure about different outputs on other devices, eg. different ARM versions, so added a dictionary which translates these strings to the correct one found in <jre_home>/lib/*.
In case a platform.machine output is unknown to the dictionary, the code will fall back to i386 like the old code and warn the user about that. A missing translation can be easily added to the dictionary to support other platforms.

WARNING: This commit is untested and was coded via github webpage.