Convert java.lang.Long to INT if < 2 ** 31 else to LONG
LONG is in Py3 merged with INT type and renamed to INT without 'L' being
present in the string representation of the number but having the full
storage capacity of LONG. On Py2 however it will cause type issues when
we convert java.lang.Long with size < 2 ** 31 into Py2 long and convert
it back. Therefore convert java.lang.Long to Py2 long *only* if it exceeds
the maximum value of INT (2 ** 31).