setup.py: Add cython to setup_requires
setup_requires is a method to add libraries that are needed during
the setup.py execution. This is picked up by easy_install and is
installed in the build directory for the build to occur.

This simplifies the installation of pyjnius from source
distributions as the cython dependency for build is managed by the
python package management toolchain.